Fixes segfault issue with triangle
Segfault was caused by use of uninitialised value inside triangle.c
which occurs when degenerate vertices are present in the input. Triangle
removes points which are coincident. My hunch is that it does not do any
reindexing of the edge vertices when this occurs which leads to the use
of values which are no longer present.